requirements.txt
flask
flask-cors
pymongo
requests==2.29.0
numpy
Pillow
keras
tensorflow
matplotlib
pandas
passlib
python-dotenv
opencv-python
python-dateutil
PyQRCode==1.2.1
pytz==2021.3
pyzbar==0.1.9
qrcode==7.3.1
six==1.16.0
tzdata==2022.1
Flask-Bcrypt
reportlab
